msm: thermal: simple: Introduce simple MSM thermal solution
When properly configured, this driver monitors temperatures from a
specified sensor and reduces the CPU's max frequency in response to
the temperature crossing specified thermal zones. The temperature sensor
use must be a specified ADC channel on a given VADC device, hence this
driver's dependency on the chipset being an MSM product.

This driver can only be configured via the device tree; it cannot be
configured at runtime. Configuration instructions can be found in the
accompanying documentation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

msm_thermal_simple: add thermal debug

Useful when stress testing the CPU so you can figure out the temps

msm: thermal: simple: Never raise max freq above unthrottled max

If the max frequency is set below the standard max, there is a
possibility of msm-thermal-simple raising the frequency in an attempt to
throttle if the target throttling zone has a higher max frequency. This
will be an issue when we unlock the max frequency setting to userspace.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

msm_thermal: ignore sensor -19

Mata has issues initializing it.
Lots and lots of messages like this in dmesg:
[   51.109389] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109419] msm_thermal:hotplug_init_cpu_offlined Unable to read
TSENS sensor:-19.
[   51.109423] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109426] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109429] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109431] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109434] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109437] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109439] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109442] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109445] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109447] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109450] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109453] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109455] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19
[   51.109458] msm_thermal:sensor_mgr_set_threshold Unable to read
temperature for zone:-19. err:-19
[   51.109461] msm_thermal:therm_get_temp Unable to read thermal zone
sensor:-19

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>

thermal: qpnp-adc-tm: Fix aborting suspend

When a low hysteresis for thresholds are set there is a
possibility that low and high thresholds are crossed at
around the same time before the client is notified. Under
such condition there could be a low and high interrupt
received before processing the request and notifying the
client. Check the return value for queued item for
notifying the client and reduce the count used to prevent
suspend if the item is already available in the global queue.

Change-Id: I4bddf6742a0686df57b6c4ce0d964579de01b027
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>
idkwhoiam322: Modified for our thermal setup.
Signed-off-by: idkwhoiam322 <snkm1999@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>

msm-thermal: omit unset nodes

Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>